# docs/openapi/booking-engine.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Booking Engine API",
    "description": "Property setup, pricing and availability for hotel distribution.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/properties/{propertyId}/rate-plans": {
      "post": {
        "tags": [
          "Rate plans"
        ],
        "summary": "Create a rate plan",
        "description": "Per-person plans price each adult count relative to the daily base rate. Derived plans take their rate from a parent plan with a percent or fixed adjustment.",
        "operationId": "CreateRatePlan",
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRatePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatePlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Rate plans"
        ],
        "summary": "List rate plans of a property",
        "operationId": "ListRatePlans",
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RatePlanResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    },
    "/api/v1/properties/{propertyId}/rate-plans/{ratePlanId}": {
      "get": {
        "tags": [
          "Rate plans"
        ],
        "summary": "Get a rate plan",
        "operationId": "GetRatePlan",
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ratePlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatePlanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Rate plans"
        ],
        "summary": "Update a rate plan",
        "description": "Code and sell mode are immutable; omitting `derived` detaches the plan from its parent.",
        "operationId": "UpdateRatePlan",
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ratePlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRatePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatePlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Rate plans"
        ],
        "summary": "Delete a rate plan",
        "operationId": "DeleteRatePlan",
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ratePlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateRatePlanRequest": {
        "required": [
          "room_type_id",
          "code",
          "name",
          "sell_mode",
          "meal_plan",
          "child_fee"
        ],
        "type": "object",
        "properties": {
          "room_type_id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "maxLength": 50,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "type": "string"
          },
          "sell_mode": {
            "$ref": "#/components/schemas/SellMode"
          },
          "meal_plan": {
            "$ref": "#/components/schemas/MealPlan"
          },
          "child_fee": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "occupancies": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/OccupancyDto"
            }
          },
          "derived": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DerivedPricingDto"
              }
            ]
          }
        }
      },
...
      "DerivedAdjustmentType": {
        "enum": [
          "percent",
          "amount"
        ]
      },
      "DerivedPricingDto": {
        "required": [
          "parent_rate_plan_id",
          "type",
          "value"
        ],
        "type": "object",
        "properties": {
          "parent_rate_plan_id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/DerivedAdjustmentType"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "MealPlan": {
        "enum": [
          "room_only",
          "breakfast",
          "half_board",
          "full_board",
          "all_inclusive"
        ]
      },
      "OccupancyDto": {
        "required": [
          "adults",
          "price_adjustment"
        ],
        "type": "object",
        "properties": {
          "adults": {
            "maximum": 20,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "price_adjustment": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RatePlanResponse": {
        "required": [
          "id",
          "room_type_id",
          "code",
          "name",
          "sell_mode",
          "meal_plan",
          "child_fee",
          "occupancies",
          "derived",
          "created_at",
          "updated_at"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "room_type_id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sell_mode": {
            "$ref": "#/components/schemas/SellMode"
          },
          "meal_plan": {
            "$ref": "#/components/schemas/MealPlan"
          },
          "child_fee": {
            "type": "number",
            "format": "double"
          },
          "occupancies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OccupancyDto"
            }
          },
          "derived": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DerivedPricingDto"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SellMode": {
        "enum": [
          "per_room",
          "per_person"
        ]
      },
...
      "UpdateRatePlanRequest": {
        "required": [
          "name",
          "meal_plan",
          "child_fee"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "type": "string"
          },
          "meal_plan": {
            "$ref": "#/components/schemas/MealPlan"
          },
          "child_fee": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "occupancies": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/OccupancyDto"
            }
          },
          "derived": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DerivedPricingDto"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Management API key.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
...
}